feat(custom-hook): isolate mutation logic from UI
